Reorganize p2p integration tests
